pyproject.toml
[project]
name = "yojanaconnect"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.3.0",
    "fastapi>=0.115.12",
    "mysqlclient>=2.2.7",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pyjwt[crypto]>=2.10.1",
    "pymysql>=1.1.1",
    "python-dotenv>=1.1.0",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.34.3",
]
